+++ apps/web/src/stories/Button/Button.stories.ts
@@ -1,21 +1,21 @@
-// 예시입니다.
 import type { Meta, StoryObj } from '@storybook/react-vite'
 
 import { fn } from 'storybook/test'
 
 import { Button } from '@/shared/ui/common/Button/Button'
 
-// More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
-  title: 'Button',
+  title: 'Components/Button/Button',
   component: Button,
   parameters: {
-    // Optional parameter to center the component in the Canv . More info: https://storybook.js.org/docs/configure/story-layout
     layout: 'centered',
+    docs: {
+      description: {
+        component: '기본 버튼 컴포넌트입니다.',
+      },
+    },
   },
-  // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
   tags: ['autodocs'],
-  // More on argTypes: https://storybook.js.org/docs/api/argtypes
   argTypes: {
     variant: {
       control: 'select',
@@ -32,14 +32,12 @@ const meta = {
     },
     disabled: { control: 'boolean' },
   },
-  // Use `fn` to spy on the onClick arg, which will appear in the actions panel once invoked: https://storybook.js.org/docs/essentials/actions#action-args
   args: { onClick: fn() },
 } satisfies Meta<typeof Button>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
-// More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
 export const Primary: Story = {
   args: {
     variant: 'primary',
