--- apps/web/src/stories/EmotionButton.stories.ts
@@ -1,43 +1,0 @@
-// 예시입니다.
-import type { Meta, StoryObj } from '@storybook/react-vite'
-
-import { fn } from 'storybook/test'
-
-import { EmotionButton } from '@/shared/ui/common/Button/EmotionButton'
-
-// More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
-const meta = {
-  title: 'EmotionButton',
-  component: EmotionButton,
-  parameters: {
-    // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
-    layout: 'centered',
-  },
-  // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
-  tags: ['autodocs'],
-  // More on argTypes: https://storybook.js.org/docs/api/argtypes
-  argTypes: {
-    emotionType: {
-      control: 'select',
-      options: ['joy', 'angry'],
-    },
-  },
-  // Use `fn` to spy on the onClick arg, which will appear in the actions panel once invoked: https://storybook.js.org/docs/essentials/actions#action-args
-  args: { onClick: fn() },
-} satisfies Meta<typeof EmotionButton>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-// More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
-export const Joy: Story = {
-  args: {
-    emotionType: 'joy',
-  },
-}
-
-export const Angry: Story = {
-  args: {
-    emotionType: 'angry',
-  },
-}
